requirements.txt
# PDF parsing
pypdf>=3.0.0

# LLM API
openai>=1.30.0
anthropic>=0.40.0
httpx>=0.27.0
tenacity>=8.2.0

# Embeddings + vector store
sentence-transformers>=2.2.0
qdrant-client>=1.7.0

# arXiv paper fetching
arxiv>=2.1.0

# Env + utilities
python-dotenv>=1.0.0

# Dev/test
pytest>=7.4.0
